==> 02-pandas-decision-trees.ipynb <==
## 1. Pandas
This tutorial is assuming that you have some basic programming skills. If you are more familiar with R than with Python, don't worry, it's not that different. First, we will have a look at the pandas package, which introduces the data frame (similar to R) as a fundamental data structure. 

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

df = pd.read_csv("iris.csv")

df

### Selecting rows / columns
Using pandas, we can easily select rows or columns in a data frame. One option is to select via name of the column:

accx = df.loc[:,"SepalLength"]
accx

Here, `:` means "choose all rows". When only a single column is selected, the result is of type `Series`, otherwise it is a data frame again. We can also access columns via their index:

acc = df.iloc[:,1:4]
acc

We can access rows in the same way. The following expression gives us a data frame consisting of the first five rows:

df.iloc[0:5,:]

These can both be combined like this:

df.loc[0:5,["SepalLength", "SepalWidth"]]

Another option is to access rows or columns via boolean expressions. For example, the following line returns a data frame that only contains rows whose `SepalLength` is smaller than 5:

df.loc[df.SepalLength < 5,:]

### Adding new values
Writing new values in a data frame is also easy. The following expression replaces all values of the column `SepalLength` that are smaller than 5 to the value 0.


df.loc[df.SepalLength < 0,"SepalLength"] = 0
df

## 2. Decision Trees in Python 
The following example trains a decision tree on the Iris data, using 2/3 as training data and 1/3 as test data.

import sklearn

X = df.iloc[:,1:5]
y = df.iloc[:,5]

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split( X, y, test_size=0.33)

from sklearn import tree
clf = tree.DecisionTreeClassifier()
clf = clf.fit(X_train, y_train)

pred = clf.predict(X_test)
print(np.mean(pred==y_test))

## 3. Task: Train your own model
Next, you can try to train a model yourself: we want you to use the Titanic passenger data (name, age, price of ticket, etc) to try to predict who will survive and who will die.

titanic = pd.read_csv("titanic.csv")

## Explain the model
Compute SHAP values of the model and plot them as a beeswarm plot.
