==> 05_feature_importance.ipynb <==
Feature importances are a way to look into why/how a model makes its decisions.

import joblib
import pandas as pd



clf = joblib.load("../clfmodel.pkl")

feature_importances = clf.feature_importances_


Let's take a quick look into the importance values and their distribution.

import matplotlib.pyplot as plt

plt.plot(sorted(feature_importances))
plt.show()

print(len(feature_importances), sum(feature_importances == 0))

As quite some features received an importance of zero, why not remove them from the embedding?

df = joblib.load("../embedded_photos_labelled.pkl")
df["filtered_embedding"] = pd.Series( entry[feature_importances != 0] for entry in df["embedding"] )
joblib.dump(df, "../embedded_photos_labelled_selected.pkl")

df = joblib.load("../embedded_photos_all_labelled.pkl")
df["filtered_embedding"] = pd.Series( entry[feature_importances != 0] for entry in df["embedding"] )
joblib.dump(df, "../embedded_photos_all_labelled_selected.pkl")
