==> 06a_dimensionality_reduction_solutions.ipynb <==
We have vector embeddings of all images and labels for some of them. Instead of simply trusting the RF output, we might want to use some methods from semi-supervised learning to check these. So we'll try to visualize the images in relation to each other using dimensionality reduction methods. But let's first load the data.

import pandas as pd
import joblib

training_vector_data = joblib.load("/home/sperlea/Desktop/project_auto_microscopy/scripts/playground/data/embedded_training_data_selected.pkl")
training_vector_data["source"] = "training"
IOW_vector_data = joblib.load("/home/sperlea/Desktop/project_auto_microscopy/scripts/playground/data/embedded_photos_labelled_selected.pkl")
IOW_vector_data["source"] = "IOW"
vector_data = pd.concat([training_vector_data, IOW_vector_data])
vector_data = vector_data.reset_index()


There are many different methods for dimensionality reduction - we'll look into PCA, MDS and T-SNE. 

import numpy as np
from sklearn.decomposition import PCA

pca = PCA(n_components=2)
dimred_vector_data = pca.fit_transform(np.vstack(vector_data["filtered_embedding"].values))

import numpy as np
from sklearn.manifold import MDS

mds = MDS(n_components=2)
dimred_vector_data = mds.fit_transform(np.vstack(vector_data["filtered_embedding"].values))

import numpy as np
from sklearn.manifold import TSNE

dimred_vector_data = TSNE(n_components=2, learning_rate='auto',
                  init='random', perplexity=3).fit_transform(np.vstack(vector_data["filtered_embedding"].values))



##And let's look at the results.

import matplotlib.pyplot as plt

fig, ax = plt.subplots() ##prepare a canvas to draw on
for type_ in set(vector_data["class"].values): ## plot one class after the other in order to have a color legend
    ax.scatter(
        dimred_vector_data[(vector_data["class"].values == type_) &
                            (vector_data["source"]  == "training"), 0],
        dimred_vector_data[(vector_data["class"].values == type_) &
                           (vector_data["source"]  == "training"), 1],
        marker='o',color = "black", s=100) ##Highlight the images from the training data using black circles

    ax.scatter(
        dimred_vector_data[vector_data["class"].values == type_, 0],
        dimred_vector_data[vector_data["class"].values == type_, 1],
        marker = 'o', label = type_) ##plot all vectors in colors

ax.legend()
plt.show()

##Let's make it easier for us to look into the data
dimred_vector_data = pd.DataFrame(dimred_vector_data)
dimred_vector_data["class"] = vector_data["class"]
dimred_vector_data["source"] = vector_data["source"]
